Add support for decoding into supplied bitmaps
On Honeycomb and higher, allow the user to provide a target bitmap for
decoding to reduce memory allocations. This is particularly useful when
using Picasso for a tile mapped application.

Using inBitmap() implicitly causes the memory cache to be skipped so that
the user has full control over when a bitmap can be reused.

The Config of the supplied bitmap is implicitly used as the Config for
decoding.